+++ src/email.ts
@@ -2,9 +2,11 @@ import { SESClient, SendEmailCommand } from "@aws-sdk/client-ses";
 
 // Optional: Import types for better type checking and autocompletion
 import { SendEmailRequest, SendEmailResponse } from "@aws-sdk/client-ses";
-import { AnyLengthString } from "aws-sdk/clients/comprehend";
+
+const awsRegion = process.env.AWS_REGION || "us-west-2";
 
 const monitoringEmail : string = process.env.MONITORING_EMAIL || "monitoring@polyverse.com";
+export const PolyverseSupportEmail : string = process.env.POLYVERSE_SUPPORT_EMAIL || `support@polyverse.com`;
 
 export async function sendMonitoringEmail(subject: string, body: string) {
 	await sendEmail(`[Boost Monitoring] ${subject}`, body, monitoringEmail);
@@ -17,11 +19,11 @@ export async function sendEmail(
 	senderEmail: string = monitoringEmail
 	): Promise<void>  {
 	if (recipientEmail === "") {
-		console.log(`SIMULATING SES EMAIL (blank recipient): ${subject}`);
+		console.log(`SIMULATING SES EMAIL (blank recipient) from ${senderEmail}: ${subject} - ${body}`);
 		return;
 	}
 
-	const client = new SESClient({ region: "us-west-2" });
+	const client = new SESClient({ region: awsRegion });
 
 	const params: SendEmailRequest = {
 		Source: senderEmail,
@@ -43,7 +45,7 @@ export async function sendEmail(
 	};
 
     if (!process.env.EMAIL_NOTIFICATIONS) {
-        console.log(`Boost Service Send Email (disabled) from ${senderEmail} to ${recipientEmail}: ${subject} ${body}`);
+        console.log(`Boost Service Send Email (disabled) from ${senderEmail} to ${recipientEmail}: ${subject} - ${body}`);
         return;
     }
 
@@ -55,8 +57,58 @@ export async function sendEmail(
 			);
 	} catch (error: any) {
 		console.error(
-			`Boost Service Send Email failed from ${senderEmail} to ${recipientEmail}: ${subject} ${body}`,
+			`Boost Service Send Email failed from ${senderEmail} to ${recipientEmail}: ${subject} - ${body}`,
 			error.stack || error);
 	}
 };
-		
+
+export async function sendHtmlEmail(
+    subject: string,
+    htmlBody: string,
+    plainTextBody: string,
+    recipientEmail: string,
+    senderEmail: string = monitoringEmail
+): Promise<void> {
+    if (recipientEmail === "") {
+        console.log(`SIMULATING SES EMAIL (blank recipient from ${senderEmail}): ${subject} - ${htmlBody}`);
+        return;
+    }
+
+    const client = new SESClient({ region: awsRegion });
+
+    const params: SendEmailRequest = {
+        Source: senderEmail,
+        Destination: {
+            ToAddresses: [recipientEmail],
+        },
+        Message: {
+            Body: {
+                Html: {
+                    Charset: "UTF-8",
+                    Data: htmlBody,
+                },
+                Text: {
+                    Charset: "UTF-8",
+                    Data: plainTextBody,
+                }
+            },
+            Subject: {
+                Charset: "UTF-8",
+                Data: subject,
+            },
+        },
+    };
+
+    if (!process.env.EMAIL_NOTIFICATIONS) {
+        console.log(`Boost Service Send Email (disabled) from ${senderEmail} to ${recipientEmail}: ${subject} ${htmlBody}`);
+        return;
+    }
+
+    try {
+        const command = new SendEmailCommand(params);
+        const response: SendEmailResponse = await client.send(command);
+        console.log(`SES Email (id:${response.MessageId}) sent to ${recipientEmail}: ${subject} - ${htmlBody}`);
+    } catch (error: any) {
+        console.error(`Boost Service Send Email failed from ${senderEmail} to ${recipientEmail}: ${subject} ${htmlBody}`, error.stack || error);
+    }
+}